pyproject.toml
[project]
name = "pyxems"
version = "0.1.0"
description = "A XML-config generator for OpenEMS"
readme = "README.md"
authors = [
    { name = "Patarimi", email = "mpqqch@gmail.com" }
]
requires-python = ">=3.12"
dependencies = []

[project.optional-dependencies]
simulate = [
    "cyclopts>=4.5.1",
    "python-dotenv>=1.2.1",
]

[build-system]
requires = ["uv_build>=0.9.15,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
